+++ logic/profession.lua
@@ -89,7 +89,7 @@ MTSL_LOGIC_PROFESSION = {
                     skill_passed_filter = false
                 end
                 -- Check availability in expansion
-                if skill_passed_filter == true and MTSL_TOOLS:ListContainsNumber(expansions, 100 * v.expansion + v.phase) == false then
+                if skill_passed_filter == true and MTSL_TOOLS:ListContainsNumber(expansions, v.expansion + v.phase) == false then
                     skill_passed_filter = false
                 end
                 -- Check availability in zone